Convert to float in CG_WideScreenMode

+++ codeJK2/cgame/cg_main.cpp
@@ -472,10 +472,10 @@ Make 2D drawing functions use widescreen or 640x480 coordinates
 */
 void CG_WideScreenMode(qboolean on) {
 	if (on) {
-		trap_MVAPI_SetVirtualScreen(cgs.screenWidth, SCREEN_HEIGHT);
+		trap_MVAPI_SetVirtualScreen(cgs.screenWidth, (float)SCREEN_HEIGHT);
 	}
 	else {
-		trap_MVAPI_SetVirtualScreen(SCREEN_WIDTH, SCREEN_HEIGHT);
+		trap_MVAPI_SetVirtualScreen((float)SCREEN_WIDTH, (float)SCREEN_HEIGHT);
 	}
 }
